Fix crash when multiple threads may exist.
Based on a crash I experienced in an app I am developing, it appears that sometimes when the app is shutting down, the disposal will not happen from the UI thread, and Dispose() on the HwndSource causes an InvalidOperationException.

By running the Dispose on the UI thread immediately, disposal happens, and no crash is seen.